+++ src/main/scala/controller/baseImp/GameMode.scala
@@ -1,4 +1,4 @@
-package controller
+package controller.baseImp
 
 import util.*
 import util.command.GameState
