+++ src/main/scala/model/gameModelComp/ModelInterface.scala
@@ -1,4 +1,4 @@
-package model
+package model.gameModelComp
 
 trait CardInterface{
     def getColor: String
