+++ src/main/scala/model/gameModelComp/baseImp/Value.scala
@@ -1,4 +1,4 @@
-package model.baseImp
+package model.gameModelComp.baseImp
 
 object Value extends Enumeration {
     type Value1 = Value
