+++ src/test/scala/controller/GameControllerSpec.scala
@@ -1,8 +1,11 @@
 package controller
 
-import model.CardInterface
-import model.FileIO.{FileIOInterface, FileIOXML}
+import controller.baseImp.{GameController, GameMode}
+import model.fileIOComp.FileIOInterface
 import model.baseImp.*
+import model.fileIOComp.baseImp.FileIOXML
+import model.gameModelComp.CardInterface
+import model.gameModelComp.baseImp.{Deck, Grid, Hand, NumberCards, Player, Suit, Value}
 import org.mockito.ArgumentMatchers.*
 import org.mockito.Mockito.*
 import org.scalatest.matchers.should.Matchers
