+++ src/test/scala/model/baseImp/NumberCardsSpec.scala
@@ -1,5 +1,6 @@
 package model.baseImp
 
+import model.gameModelComp.baseImp.{NumberCards, Suit, Value}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
