+++ src/test/scala/model/baseImp/SuitSpec.scala
@@ -1,5 +1,6 @@
 package model.baseImp
 
+import model.gameModelComp.baseImp.Suit
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
