+++ src/test/scala/model/baseImp/ValueSpec.scala
@@ -1,5 +1,6 @@
 package model.baseImp
 
+import model.gameModelComp.baseImp.Value
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
